Add validations and error handling to logger

+++ Application/wwwroot/Js.js
@@ -1,9 +1,6 @@
 ﻿
 $(document).ready(function () {
 
-
-
-
     const uri = 'ViewDevices';
 
     $("#showAllBtn").click(function () {
@@ -29,6 +26,10 @@ $(document).ready(function () {
 
     $("#showMaxMinBtn").click(function () {
         var minField = $("#showMaxMin")[0].valueAsNumber;
+        if (!minField) {
+            alert("Empty Input. Fill it, and try again");
+            return;
+        }
         var tmpUri = uri + '/minField?minField=' + minField;
         $.getJSON(tmpUri)
             .done(function (data) {
@@ -41,6 +42,10 @@ $(document).ready(function () {
 
     $("#showByTypeBtn").click(function () {
         var type = $("#showByType")[0].value;
+        if (!type) {
+            alert("Empty Input. Fill it, and try again");
+            return;
+        }
         var tmpUri = uri + '/type?type=' + type;
         $.getJSON(tmpUri)
             .done(function (data) {
@@ -53,6 +58,10 @@ $(document).ready(function () {
 
     $("#deleteBtn").click(function () {
         var id = $("#devideIdDelete")[0].valueAsNumber;
+        if (!id) {
+            alert("Empty Input. Fill it, and try again");
+            return;
+        }
         $.ajax({
             url: uri + '?id=' + id,
             type: 'DELETE',
@@ -68,6 +77,11 @@ $(document).ready(function () {
         var aerialLine = $("#aerialLine")[0].valueAsNumber;
         var degrees = $("#degrees")[0].valueAsNumber;
         var engle = $("#angle")[0].valueAsNumber;
+
+        if (!TypeName || !meters || !aerialLine || !degrees || !engle) {
+            alert("Empty Input. Fill it, and try again");
+            return;
+        }
 
         var data = {
             Range: { AerialLine: aerialLine, Meters: meters },
@@ -86,48 +100,4 @@ $(document).ready(function () {
             }
         });
     });
-
-    
-
-
-    function _displayItems(data) {
-        const tBody = document.getElementById('todos');
-        tBody.innerHTML = '';
-
-        _displayCount(data.length);
-
-        const button = document.createElement('button');
-
-        data.forEach(item => {
-            let isCompleteCheckbox = document.createElement('input');
-            isCompleteCheckbox.type = 'checkbox';
-            isCompleteCheckbox.disabled = true;
-            isCompleteCheckbox.checked = item.isComplete;
-
-            let editButton = button.cloneNode(false);
-            editButton.innerText = 'Edit';
-            editButton.setAttribute('onclick', `displayEditForm(${item.id})`);
-
-            let deleteButton = button.cloneNode(false);
-            deleteButton.innerText = 'Delete';
-            deleteButton.setAttribute('onclick', `deleteItem(${item.id})`);
-
-            let tr = tBody.insertRow();
-
-            let td1 = tr.insertCell(0);
-            td1.appendChild(isCompleteCheckbox);
-
-            let td2 = tr.insertCell(1);
-            let textNode = document.createTextNode(item.name);
-            td2.appendChild(textNode);
-
-            let td3 = tr.insertCell(2);
-            td3.appendChild(editButton);
-
-            let td4 = tr.insertCell(3);
-            td4.appendChild(deleteButton);
-        });
-
-        todos = data;
-    }
 })
